Local tests work

+++ src/GCE_helpers.py
@@ -1,6 +1,7 @@
 import configparser
 import pickle
 from google.cloud import storage
+import os
 
 class GCE_control:
     def __init__(self):
@@ -15,38 +16,41 @@ class GCE_control:
         """
         If we are running inside a GCE VM, kill it.
         """
-        print("Trying to kill VM")
-        # from https://stackoverflow.com/a/52811140/10581449
-        import json
-        import logging
-        import requests
+        if not 'AM_I_IN_A_DOCKER_CONTAINER' in os.environ:
+            pass
+        else:
+            self.send_email_update("Finished script - trying to kill VM")
+            # from https://stackoverflow.com/a/52811140/10581449
+            import json
+            import logging
+            import requests
 
-        # get the token
-        r = json.loads(
-            requests.get("http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/token",
-                        headers={"Metadata-Flavor": "Google"})
-                .text)
+            # get the token
+            r = json.loads(
+                requests.get("http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/token",
+                            headers={"Metadata-Flavor": "Google"})
+                    .text)
 
-        token = r["access_token"]
+            token = r["access_token"]
 
-        # get instance metadata
-        # based on https://cloud.google.com/compute/docs/storing-retrieving-metadata
-        project_id = requests.get("http://metadata.google.internal/computeMetadata/v1/project/project-id",
+            # get instance metadata
+            # based on https://cloud.google.com/compute/docs/storing-retrieving-metadata
+            project_id = requests.get("http://metadata.google.internal/computeMetadata/v1/project/project-id",
+                                    headers={"Metadata-Flavor": "Google"}).text
+
+            name = requests.get("http://metadata.google.internal/computeMetadata/v1/instance/name",
                                 headers={"Metadata-Flavor": "Google"}).text
 
-        name = requests.get("http://metadata.google.internal/computeMetadata/v1/instance/name",
-                            headers={"Metadata-Flavor": "Google"}).text
+            zone_long = requests.get("http://metadata.google.internal/computeMetadata/v1/instance/zone",
+                                    headers={"Metadata-Flavor": "Google"}).text
+            zone = zone_long.split("/")[-1]
 
-        zone_long = requests.get("http://metadata.google.internal/computeMetadata/v1/instance/zone",
-                                headers={"Metadata-Flavor": "Google"}).text
-        zone = zone_long.split("/")[-1]
+            # shut ourselves down
+            logging.info("Calling API to delete this VM, {zone}/{name}".format(zone=zone, name=name))
 
-        # shut ourselves down
-        logging.info("Calling API to delete this VM, {zone}/{name}".format(zone=zone, name=name))
-
-        requests.delete("https://www.googleapis.com/compute/v1/projects/{project_id}/zones/{zone}/instances/{name}"
-                        .format(project_id=project_id, zone=zone, name=name),
-                        headers={"Authorization": "Bearer {token}".format(token=token)})
+            requests.delete("https://www.googleapis.com/compute/v1/projects/{project_id}/zones/{zone}/instances/{name}"
+                            .format(project_id=project_id, zone=zone, name=name),
+                            headers={"Authorization": "Bearer {token}".format(token=token)})
 
     def send_email_update(self, message, subject = "Update on GCE Python Script"):
         import smtplib
@@ -65,14 +69,14 @@ class GCE_control:
         if (bucket.exists()):
             return storage_client.get_bucket(self.BUCKET_NAME)
         else:    
-            return storage_client.create_bucket(bucket, location="us")
+            return storage_client.create_bucket(bucket, location="us", project=self.PROJECT_NAME)
 
     def save_output(self, out, filename = None):
-        bucket = self.create_or_get_bucket(self.BUCKET_NAME)
+        bucket = self.create_or_get_bucket()
 
         self.output_counter += 1    
         if filename is None:
-            filename("script_output_" + self.output_counter + ".pkl")
+            filename = "script_output_" + str(self.output_counter) + ".pkl"
 
         if not filename[-4:] == ".pkl":
             filename += ".pkl"
